Create HistoryProvider.loadNewTracks()
The Spotify Web API does not allow getting track history with both after
and before cursors set. after is used for getting newer history entries
and before is used for older history entries. This requires two ways of
updating the HistoryProvider track cache, so the logic was moved out of
the getChildren() function and into loadNewTracks(). getChildren() now
just returns the track cache.